requirements.txt
discord.py
gtts
ffmpeg-python
